+++ numba_dpex/core/overloads/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-FileCopyrightText: 2020 - 2024 Intel Corporation
+#
+# SPDX-License-Identifier: Apache-2.0
